Add more guards
