+++ src/hiperhealth/pipeline/runner.py
@@ -11,12 +11,11 @@ from typing import TYPE_CHECKING, Any
 from hiperhealth.pipeline.context import AuditEntry, PipelineContext
 from hiperhealth.pipeline.session import Inquiry
 from hiperhealth.pipeline.skill import Skill
+from hiperhealth.utils import _scrub_sensitive_data
 
 if TYPE_CHECKING:
     from hiperhealth.pipeline.registry import SkillRegistry
     from hiperhealth.pipeline.session import Session
-
-_SENSITIVE_PATTERNS = {'key', 'token', 'secret', 'password', 'pwd'}
 
 
 class StageRunner:
@@ -392,48 +391,6 @@ class StageRunner:
         )
         session.update_from_context(stage, ctx)
         return session
-
-
-def _scrub_sensitive_data(data: Any) -> Any:
-    """
-    title: Recursively mask sensitive fields in dictionaries or objects.
-    parameters:
-      data:
-        type: Any
-    returns:
-      type: Any
-    """
-    if isinstance(data, dict):
-        return {
-            k: '********' if _is_sensitive_key(k) else _scrub_sensitive_data(v)
-            for k, v in data.items()
-        }
-    if isinstance(data, (list, tuple)):
-        return [_scrub_sensitive_data(i) for i in data]
-
-    # Handle LLMSettings specifically if it appears in kwargs
-    try:
-        from hiperhealth.llm import LLMSettings
-
-        if isinstance(data, LLMSettings):
-            return 'LLMSettings(masked)'
-    except ImportError:
-        pass
-
-    return data
-
-
-def _is_sensitive_key(key: str) -> bool:
-    """
-    title: Check if a key name suggests it contains sensitive information.
-    parameters:
-      key:
-        type: str
-    returns:
-      type: bool
-    """
-    name = key.lower()
-    return any(pattern in name for pattern in _SENSITIVE_PATTERNS)
 
 
 __all__ = ['StageRunner']
